pyproject.toml
[tool.poetry]
name = "oz-01-main-004-be"
version = "0.1.0"
description = ""
authors = ["SwanyCastle <seunghwan712@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
django-environ = "^0.11.2"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
gunicorn = "^22.0.0"
requests = "^2.31.0"
boto3 = "^1.34.98"
django-storages = "^1.14.3"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
ignore_missing_imports = true
python_version = 3.12
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
mypy = "^1.10.0"
django-stubs = "^5.0.0"
djangorestframework-stubs = "^3.15.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
